chore: move extension-updater and extension-catalog as subfolders of new extension folder
related to https://github.com/podman-desktop/podman-desktop/pull/10623
