+++ packages/main/src/plugin/extension/catalog/extensions-catalog-api.ts
