+++ packages/main/src/plugin/extension/catalog/extensions-catalog-settings.ts
