+++ packages/main/src/plugin/extension/updater/extensions-updater-settings.ts
