+++ packages/main/src/plugin/extension/updater/extensions-updater.spec.ts
@@ -19,14 +19,14 @@
 import { app } from 'electron';
 import { afterEach, beforeEach, expect, test, vi } from 'vitest';
 
+import type { ConfigurationRegistry } from '/@/plugin/configuration-registry.js';
+import type { ExtensionsCatalog } from '/@/plugin/extension/catalog/extensions-catalog.js';
+import type { CatalogExtension } from '/@/plugin/extension/catalog/extensions-catalog-api.js';
 import type { ExtensionLoader } from '/@/plugin/extension/extension-loader.js';
+import type { ExtensionInstaller } from '/@/plugin/install/extension-installer.js';
+import type { Telemetry } from '/@/plugin/telemetry/telemetry.js';
 import type { ExtensionInfo } from '/@api/extension-info.js';
 
-import type { ConfigurationRegistry } from '../configuration-registry.js';
-import type { ExtensionsCatalog } from '../extensions-catalog/extensions-catalog.js';
-import type { CatalogExtension } from '../extensions-catalog/extensions-catalog-api.js';
-import type { ExtensionInstaller } from '../install/extension-installer.js';
-import type { Telemetry } from '../telemetry/telemetry.js';
 import { ExtensionsUpdater } from './extensions-updater.js';
 
 vi.mock('electron', () => ({
