+++ packages/main/src/plugin/featured/featured.spec.ts
@@ -18,9 +18,9 @@
 
 import { afterEach, beforeAll, beforeEach, expect, test, vi } from 'vitest';
 
+import type { ExtensionsCatalog } from '/@/plugin/extension/catalog/extensions-catalog.js';
 import type { ExtensionLoader } from '/@/plugin/extension/extension-loader.js';
 
-import type { ExtensionsCatalog } from '../extensions-catalog/extensions-catalog.js';
 import { Featured } from './featured.js';
 
 let featured: Featured;
